move user class to model folder add validation to the user model change route Recive user insted of parameters
