+++ Sat.Recruitment.Api/Controllers/UsersController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-
+using Sat.Recruitment.Api.Models;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -15,7 +15,7 @@ namespace Sat.Recruitment.Api.Controllers
     }
 
     [ApiController]
-    [Route("[controller]")]
+    [Route("api/[controller]")]
     public partial class UsersController : ControllerBase
     {
 
@@ -25,63 +25,46 @@ namespace Sat.Recruitment.Api.Controllers
         }
 
         [HttpPost]
-        [Route("/create-user")]
-        public async Task<Result> CreateUser(string name, string email, string address, string phone, string userType, string money)
+        public async Task<IActionResult> CreateUser([FromBody] User newUser)
         {
-            var errors = "";
-
-            ValidateErrors(name, email, address, phone, ref errors);
-
-            if (errors != null && errors != "")
-                return new Result()
-                {
-                    IsSuccess = false,
-                    Errors = errors
-                };
-
-            var newUser = new User
+            if (!ModelState.IsValid)
             {
-                Name = name,
-                Email = email,
-                Address = address,
-                Phone = phone,
-                UserType = userType,
-                Money = decimal.Parse(money)
-            };
+                return BadRequest(ModelState);
+            }
 
             if (newUser.UserType == "Normal")
             {
-                if (decimal.Parse(money) > 100)
+                if (newUser.Money > 100)
                 {
                     var percentage = Convert.ToDecimal(0.12);
                     //If new user is normal and has more than USD100
-                    var gif = decimal.Parse(money) * percentage;
+                    var gif = newUser.Money * percentage;
                     newUser.Money = newUser.Money + gif;
                 }
-                if (decimal.Parse(money) < 100)
+                if (newUser.Money < 100)
                 {
-                    if (decimal.Parse(money) > 10)
+                    if (newUser.Money > 10)
                     {
                         var percentage = Convert.ToDecimal(0.8);
-                        var gif = decimal.Parse(money) * percentage;
+                        var gif = newUser.Money * percentage;
                         newUser.Money = newUser.Money + gif;
                     }
                 }
             }
             if (newUser.UserType == "SuperUser")
             {
-                if (decimal.Parse(money) > 100)
+                if (newUser.Money > 100)
                 {
                     var percentage = Convert.ToDecimal(0.20);
-                    var gif = decimal.Parse(money) * percentage;
+                    var gif = newUser.Money * percentage;
                     newUser.Money = newUser.Money + gif;
                 }
             }
             if (newUser.UserType == "Premium")
             {
-                if (decimal.Parse(money) > 100)
+                if (newUser.Money > 100)
                 {
-                    var gif = decimal.Parse(money) * 2;
+                    var gif = newUser.Money * 2;
                     newUser.Money = newUser.Money + gif;
                 }
             }
@@ -139,38 +122,39 @@ namespace Sat.Recruitment.Api.Controllers
                 {
                     Debug.WriteLine("User Created");
 
-                    return new Result()
+                    return Ok(new Result
                     {
                         IsSuccess = true,
                         Errors = "User Created"
-                    };
+                    });
                 }
                 else
                 {
                     Debug.WriteLine("The user is duplicated");
 
-                    return new Result()
+                    return Ok(new Result
                     {
                         IsSuccess = false,
                         Errors = "The user is duplicated"
-                    };
+                    });
                 }
             }
             catch
             {
                 Debug.WriteLine("The user is duplicated");
-                return new Result()
+              
+                return Ok(new Result
                 {
                     IsSuccess = false,
                     Errors = "The user is duplicated"
-                };
+                });
             }
 
-            return new Result()
+            return Ok(new Result
             {
                 IsSuccess = true,
                 Errors = "User Created"
-            };
+            });
         }
 
         //Validate errors
@@ -185,18 +169,10 @@ namespace Sat.Recruitment.Api.Controllers
             if (address == null)
                 //Validate if Address is null
                 errors = errors + " The address is required";
+            //Validate if Phone is null
             if (phone == null)
-                //Validate if Phone is null
                 errors = errors + " The phone is required";
         }
     }
-    public class User
-    {
-        public string Name { get; set; }
-        public string Email { get; set; }
-        public string Address { get; set; }
-        public string Phone { get; set; }
-        public string UserType { get; set; }
-        public decimal Money { get; set; }
-    }
+
 }
